backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
python-jose[cryptography]
bcrypt==4.0.1
scikit-learn
pandas
numpy
websockets
supabase
psycopg2-binary
